+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to `homekit-ratgdo` will be documented in this file. This pr
 
 ### What's Changed
 
-* New feature... allow setting WiFi SSID.
+* New feature... allow setting WiFi SSID (<https://github.com/ratgdo/homekit-ratgdo/issues/224>).
 * New feature... enable IRAM heap to increase available memory.
+* New feature... move all user config settings into single file, improves boot time by ~13 seconds (<https://github.com/ratgdo/homekit-ratgdo/issues/241>).
 * Bugfix... ensure that network hostname is RFC952 compliant (e.g. no spaces).
 
 ### Known Issues